Do not rename simple keywords
Although this could sometimes be useful, in most cases it's too
dangerous to allow since it's easy to forget that keywords are global in
scope.

In a keyword destructuring form, it was random whether the scoped symbol
or the keyword would be renamed. This change also made it so the symbol
was always preferred.