Unescape JAR URI.
It appears that certain clients send to clojure-lsp, the URI (following
the scheme) encoded - namely the characters `:` and `!` which are
encoded as `%3A` and `%21` respectively.

This encoding unfortunately prevents the JAR from being opened for
returning back to the client for display.

By using `java.net.Decoder/decode`, the URI is unescaped back to its
original form, thus allowing the JAR to be opened successfully.

Fixes #492

-=david=-