Extend move-coll-entry to elementwise movements
Originally move-coll-entry moved elements in vectors and maps in a
pairwise fashion. This extends move-coll-entry so that it will move
elementwise instead.  That is, it will move a single element in some
situations but a pair of elements in others. It tries to be smart about
which situation it is being called in. It will move pairwise in maps and
vectors that establish bindings, and elementwise in sets, lists and
regular vectors.

As before, it does its best to preserve comment affiliation. Comments
will be kept near to the element as it moves, including comments that
appear on the lines before an element or a comment at the end of the
same line as the element.