When parsing, allow zloc to be whitespace or comment
Before this change, when parsing a file's contents into a rewrite-clj
zipper, the returned zloc was never a comment or whitespace node. That
is, it was always sexpr-able. If the parser was invoked while the cursor
was on a comment or whitespace node, the zipper was moved to the z/next
sexp following the comment or whitespace, if any.

After this change, the zloc's node may be whitespace or a comment. This
preserves the cursor's position more accurately, and in turn, allows
move-coll-entry to move an entry affiliated with a comment or
whitespace.

When I was initially researching this change I expected many test
failures. I figured many refactorings would expect to start at an
sexpr-able node. As the diff shows, there were actually only a few
failing tests, and a correspondingly small number of implementation
fixes:

1. The tests for the parser itself needed to be adjusted slightly
2. When starting on whitespace, the thread-all refactorings expected to
   be on a non-whitespace/comment node. This behavior was easily
   restored.
3. When starting on a comment, the extract-function, introduce-let, and
   move-to-let refactorings all needed to be adjusted to operate on a
   real node, not on the comment.