Properly position cursor after moving multiline elements
Before this change, when swapping elements the cursor sometimes ended up
in the wrong place. This happened when moving down and when one element
had more lines than the other. Now the cursor is always moved with the
element, ending up at its beginning.

This is particularly important when invoking move up or down repeatedly.
If the cursor jumps around between or within elements, repeated
invocations can end up moving the wrong thing.