When parsing, allow zloc to be whitespace or comment (#716)
* When parsing, allow zloc to be whitespace or comment

Before this change, when parsing a file's contents into a rewrite-clj
zipper, the returned zloc was never a comment or whitespace node. That
is, it was always sexpr-able. If the parser was invoked while the cursor
was on a comment or whitespace node, the zipper was moved to the z/next
sexp following the comment or whitespace, if any.

After this change, the zloc's node may be whitespace or a comment. This
preserves the cursor's position more accurately, and in turn, allows
move-coll-entry to move an entry affiliated with a comment or
whitespace.

When I was initially researching this change I expected many test
failures. I figured many refactorings would expect to start at an
sexpr-able node. As the diff shows, there were actually only a few
failing tests, and a correspondingly small number of implementation
fixes:

1. The tests for the parser itself needed to be adjusted slightly
2. When starting on whitespace, the thread-all refactorings expected to
   be on a non-whitespace/comment node. This behavior was easily
   restored.
3. When starting on a comment, the extract-function, introduce-let, and
   move-to-let refactorings all needed to be adjusted to operate on a
   real node, not on the comment.

* Avoid errors when completion is invoked on comment

* Add instructions for developing clojure-lsp with Emacs

* Abort refactorings when no appropriate zloc can be found

* Avoid errors when completing unqualified keywords

* Lint whitespace

* Avoid errors when extracting function at end of file

* Avoid errors when suppressing diagnostic outside form

* Improve docs about developing with Emacs

* Lint whitespace

* Pass transform tests through parser

This passes most of the transform tests through the parser, which is how
the transforms are actually invoked. This keeps the tests closer to the
true behavior of the library. It also minimizes the amount of zipper
manipulation needed in the tests.

* Lint ns

* Pass missing libspec tests through parser

* Pass edit tests through parser

* Tidy test helpers

* Avoid errors when cycling non-collections

* Avoid errors when inlining non-symbols

* Lint fix

* Avoid errors when adding missing libspec from invalid location

* Simplify add-require-suggestion and document a bug in its usage

* Avoid errors when finding references of unknown autoresolved keywords

* Lint whitespace

* Lint whitespace

* Lint unused private var

* Re-record move-coll-entry gif

* Use test helper

* Use other test helper

Co-authored-by: Eric Dallo <ericdallo06@hotmail.com>