Improve performance of core completions
This commit improves completion performance when there are a large
number of symbols in clojure.core or cljs.core that are completions.

The fix is to change a quadratic time query into a linear time query. In
particular, let the number of elements in the analysis db be M and the
number of symbols defined in clojure.core be N. Creating completions for
all N symbols took O(M*N) time, because we ran the analysis query once
for each symbol. Now we query the analysis only once, changing run time
to O(M+N).

On my machine, when in clojure-lsp/feature/completion/completion.clj
(i.e. a 500 line file in a medium-to-large project), completing from a
whitespace node used to take about 1000ms. Now it takes about 100ms.

This wasn't noticed before because usually N is small—i.e. there aren't
many symbols in clojure.core that are completions. However, when
completion is invoked when the cursor isn't on a symbol, every symbol in
clojure.core is a completion.

This performance problem is probably the true underlying cause of
[#750][750]. That bug was patched by [#756][756] because we didn't know
the underlying cause at the time. After this patch is merged, it may be
possible to revert 756 without any detriment to performance. On the
other hand, while completing within a comment may no longer be _slow_,
it's still probably not very _useful_. So, I'd vote for leaving 756 in
place.

This bug was diagnosed by wrapping the main parts of
clojure-lsp.feature.completion/completion in
[metrics-clojure][metrics-clojure] timers, and bisecting down to slow
parts of the code. That process hasn't unearthed any other major
performance problems. I suspect that snippets are slower than they need
to be because we unnecessarily re-parse the file. I have a fix for that,
but it's in a different branch with several other changes. I'll submit
that separately. All that said, the completion code handles several
contexts, and I didn't performance test them all. There may be gains to
had in other situations.

[750]: https://github.com/clojure-lsp/clojure-lsp/issues/750
[756]: https://github.com/clojure-lsp/clojure-lsp/pull/756
[metrics-clojure]: https://metrics-clojure.readthedocs.io/en/latest/