Ignore kondo config nested under the project root (#765)
As discussed in this [thread][slack-thread], sometimes users would see
lint findings that didn't match their .clj-kondo config. This happend
both on the command line and in editors.

It was determined that if a user had, for example, both a .clj-kondo
directory and a lib/.clj-kondo directory, the settings in the second
would take precedence when clojure-lsp linted an individual file in
lib/.

The root config is used when invoking `clj-kondo` on the command line,
leading to different results depending on which tool linted the file.

We think that users never intend to have a nested config directory. If
they do, they should delete it, perhaps after merging its settings into
the root config directory.

But, to prevent confusion, this patch makes clojure-lsp ignore nested
config dirs when linting individual files.

[slack-thread]: https://app.slack.com/client/T03RZGPFR/CPABC1H61/thread/CPABC1H61-1644946235.338589