Try to expose errors on client side of integration tests
While researching [flaky integration tests][flaky-tests] we noticed that
sometimes a test will [time out][time-out] because the server has sent a
response, but the client fails to read it. No exceptions are logged, the
tests just hang waiting for the client to do something.

The client runs in a separate thread, looping over and parsing the
server's output.

This commit removes an exception handler in the client loop. The handler
was using an IOException as a signal that the server's output stream had
closed, and that therefore the client could close too. We now use a
different mechanism of detecting that the server has closed.

Though this is speculative, we hope is that this will expose unexpected
IOExceptions, which could have caused the client to stop responding.

[flaky-tests]: https://github.com/clojure-lsp/clojure-lsp/issues/704
[time-out]: https://github.com/clojure-lsp/clojure-lsp/pull/757#issuecomment-1035767474