Try different approach to stopping client loop
CI is printing IOExceptions that don't seem to cause any problems, but
which clutter the logs. This is hard to fix because I can't reproduce it
locally.

I thought readLine would only ever return nil, and that we could use
that as reliable indication of when the client should be closed.
Sometimes it does return nil, but sometimes it throws.

Perhaps we can avoid the problem by properly canceling the thread that
the client loop runs in. Before, we were closing an async channel, which
did nothing to actually stop the thread. This patch changes things to
actually cancel the thread.

On the other hand, there's a good chance that closing the thread will
simply cause a different kind of IOException.

The client loop is essentially a server, waiting for and responding to
requests. It's tricky to write a proper server, and seems like overkill
for a test helper. So, I'm just trying to find a middle ground.

I suspect that we may also need to sprinkle `with-open` throughout this
code to properly clean up resources.