Actually make assertion in test
If we end a test with a message from the server to the client, but don't
make any assertions that the client receives that message, the tests can
close the client in the middle of reading the message. This leads to
IOExceptions, which clutter the logs. This adds an assertion, which all
tests should have at least one of anyway.