Try to expose errors on client side of integration tests (#769)
* Try to expose errors on client side of integration tests

While researching [flaky integration tests][flaky-tests] we noticed that
sometimes a test will [time out][time-out] because the server has sent a
response, but the client fails to read it. No exceptions are logged, the
tests just hang waiting for the client to do something.

The client runs in a separate thread, looping over and parsing the
server's output.

This commit removes an exception handler in the client loop. The handler
was using an IOException as a signal that the server's output stream had
closed, and that therefore the client could close too. We now use a
different mechanism of detecting that the server has closed.

Though this is speculative, we hope is that this will expose unexpected
IOExceptions, which could have caused the client to stop responding.

[flaky-tests]: https://github.com/clojure-lsp/clojure-lsp/issues/704
[time-out]: https://github.com/clojure-lsp/clojure-lsp/pull/757#issuecomment-1035767474

* Try different approach to stopping client loop

CI is printing IOExceptions that don't seem to cause any problems, but
which clutter the logs. This is hard to fix because I can't reproduce it
locally.

I thought readLine would only ever return nil, and that we could use
that as reliable indication of when the client should be closed.
Sometimes it does return nil, but sometimes it throws.

Perhaps we can avoid the problem by properly canceling the thread that
the client loop runs in. Before, we were closing an async channel, which
did nothing to actually stop the thread. This patch changes things to
actually cancel the thread.

On the other hand, there's a good chance that closing the thread will
simply cause a different kind of IOException.

The client loop is essentially a server, waiting for and responding to
requests. It's tricky to write a proper server, and seems like overkill
for a test helper. So, I'm just trying to find a middle ground.

I suspect that we may also need to sprinkle `with-open` throughout this
code to properly clean up resources.

* Assign IDS to integration test clients to track lifetime

* More careful test cleanup, with more logging

* Fix indentation, rename vars

* Actually print stacktrace when client dies

The prior experiments assumed that an exception thrown in a thread would
be printed. That was true, but something must have changed—perhaps the
switch from async/thread to future. In any case, this restores tthe
printing of the stacktrace.

* Simplify logging; log more types of errors

* Re-run tests

* Re-run tests

* Actually make assertion in test

If we end a test with a message from the server to the client, but don't
make any assertions that the client receives that message, the tests can
close the client in the middle of reading the message. This leads to
IOExceptions, which clutter the logs. This adds an assertion, which all
tests should have at least one of anyway.

* Fix whitespace

* Re-run integration tests

* Re-run integration tests