Tidy require/import code actions and refactorings (#772)
* Tidy require/import code actions and refactorings

The aim of this commit is not to change any behavior but to simplify and
clarify several pieces of code.

First, in add-missing-libspec and code-actions it reorganizes the code
related to generating actions. To me, this code is much more succinct
now, though of course that's subjective.

Second, in the parser, it separates parsing from initial cursor
positioning. When these two responsibilities are separate, they are
easier to compose.

Related to this separation, there's a small performance improvement for
code action generation. We now avoid parsing the code once per
combination of diagnostic and possible resolution, and instead parse
once, repositioning the cursor as necessary.

There's also a small performance improvement in snippet generation,
where we can avoid parsing the file a second time. This fix was
mentioned in [#764][764].

The parser refactoring exposed the fact that parser/safe-zloc-of-string
was actually only _mostly_ safe. It recovered from several errors
related to parsing incomplete Clojure, but would still throw if the
string was invalid Clojure. Its name has been updated, and a new version
added that will return nil for invalid Clojure. This safe version should
be used in more places, with the calling code handling nils instead of
relying on exceptions to be thrown. This additional cleanup hasn't been
completed.

[764]: https://github.com/clojure-lsp/clojure-lsp/pull/764

* Avoid extra parse in code actions

Co-authored-by: Eric Dallo <ericdallo06@hotmail.com>