Try to avoid infinite loop when analyzing several files
This is another attempt at fixing
https://github.com/clojure-lsp/clojure-lsp/issues/796. The basic idea is
that this avoids a situation where when two files are updated
simultaneously, the first one finishes and sets :processing-changes to
false before the second one is actually done processing changes. This
creates conditions where it is more likely that the app will get into an
infinite loop. See that issue for details about what those conditions
are.