Fixes #837 (#838)
* Fixes #837

A number of logic changes to what gets suggested, the principle is not
to introduce conflicting aliases and to reuse existing ones as much as
possible.

Combines the logic for add missing libspec and find suggestions
Removed add missing libspec from code actions because it's redundant
when a user chooses.

* Fix linting

Co-authored-by: case <me@case.codes>