Exclude declare and defmethod from var references
This resolves https://github.com/clojure-lsp/clojure-lsp/issues/840 and
https://github.com/clojure-lsp/clojure-lsp/issues/841.

Before this, `declare` and `def/defn` counted in each other's usages,
and so they were never marked as unused-public-vars, even if they had no
real usages. Similarly, `defmethod` counted for `defmulti`'s usages, so
a `defmulti` with at least one `defmethod` was never marked as an
unused-public-var, even if it didn't have any real usages.

This also changes the code lens reference counts.

In the case of defmulti, it means that defmethods are no longer included
when invoking find-references (a.k.a. go to references). You can still
get from a defmulti to its defmethods with find-implementations. We may
add defmethods back to find-references at some point, at the same time
as we add protocol method implementations.