Optimize unused public var (#861)
* Research optimizations for finding unused public vars

The baseline algorithm for finding unused public vars has the following
performance characteristics.

|               |   time | std-dev |
|---------------|-------:|--------:|
| large file    |   98ms |     1ms |
| kw-heavy file |  249ms |    11ms |
| whole project | 1220ms |    11ms |

This was run on clojure-lsp itself, using queries.clj as a large file
and coercer.clj as a file with many keyword definitions.

This commit researches several faster algorithms. Let's focus on the
last algorithm, "v5". Its performance is significantly better:

|               | time | std-dev |
|---------------|-----:|--------:|
| large file    | 13ms |  0.20ms |
| kw-heavy file | 13ms |  0.20ms |
| whole project | 24ms |  0.16ms |

Impressively, and unlike the baseline algorithm, this performance is
achieved without any parallelization.

To compare the algorithms, first consider the baseline algorithm. First
it loops through the project analysis once to find var definitions and
once to find keyword definitions. Then, for each definition, it loops
through the analysis once again, looking for any usages. This has
quadratic performance, based on the size of the project analysis (# of
definitions * # of usages).

The "v5" algorithm is linear in the size of the project analysis (# of
definitions + # of usages).  Instead of looking for usages of one
definition at a time, first it looks for all usages. It loops once
through the project analysis to find all var usages, and once for all
keyword usages. Then it loops once for all var definitions, using set
logic to remove any that have usages, and another time for keyword
definitions.

This is the major optimization, converting an O(N*N) algorithm to O(N).
There are also some smaller optimizations to constrain the number of var
usages that are inspected, reducing the size of N. This is particularly
helpful during single-file analysis, when we can ignore usages of other
namespaces.

Since whole-project linting is not much slower than single-file linting,
perhaps it would be simpler to re-lint the full project after every file
change. Then we wouldn't need to re-lint the files that reference the
first file. I'll plan to test this on larger projects. Even if it looks
like it will work, it may belong in a separate PR.

The next step is to integrate the "v5" algorithm into the real code. I
plan to move what I can to the queries namespace. Any feedback at this
time is welcome.

* Improve performance of unused-public-vars linting

See [33b3f350f][analysis-commit] for an analysis of why and to what
extent this implementation is faster. This commit removes the rich
comment where that analysis was done.

[analysis-commit]: https://github.com/mainej/clojure-lsp/commit/33b3f350f4517e243b9a8483c47a95449855cc9b

* Introduce var; shorten definitions

* Fix to store findings, not their elements

* Add changelog; delegate to identical method

* More descriptively report linting and waiting time