Namespace suggestion regessions and improvements
There are number of decisions that go into ns suggestions and there's a
lot of guessing as to what the user might be intending.

Consider `clojure.data.xml` that the user is trying to require.
Question if what the user types is their intended alias or a search for
a namespace, or something in between.

* `xml` feels like intended alias.
* `d.xml` feels like it could be intended.
* `c.d.x` feels like a search.
* `d.x` feels like a search or do they intend `d.xml`?

We split the given alias and possible ns into segments and follow this
logic:
1. if only a single segment is given, suggest the best segments from the
   ns (first available alias of `xml` and `data.xml`)
2. if the last given segment matches the last ns segment, use given
   alias (`d.xml` => `[clojure.data.xml :as d.xml]`)
3. if the segment counts are the same, suggest best and a plain ns
   require (`[clojure.data.xml :as xml]` and `[clojure.data.xml]`)
4. otherwise, expand the given segment to the last ns segment (`d.x` =>
   `d.xml`)

I believe there is room here for user setting to control some of this.
A setting that follows this logic by default or always pick best, given, or plain-ns could be added in the future.
I believe there's also room for the logic to only accept given prefixes
as well. ex: If your project requires `[models.* :as m.*]` everywhere only take given
if it starts with the known alias prefix `m.` otherwise fallback to best segments.

Since a given alias could match hundreds of namespaces it's important to
be restrictive on suggestions.