More quickly identify changed files in notify-references
This removes several nested loops from the algorithm that identifies
changed incoming references to or outgoing references from a file. In
clojure2d.color, this reduces the time finding associated files from
over 30s to around 0.2s.

We haven't been logging the main body of work of notify-references, and
it happens async, so it's been out-of-sight and out-of-mind. This also
adds duration logging for notify-references. That should make it easier
to pinpoint as a performance problem later.

There are more optimizations to be done in find-references, but this is
a good start.