Optimize finding ns def by filename (#933)
* Optimize finding ns def by filename

Reduces CPU usage of refresh-test-tree and probably code-lens by
querying fewer elements. More noticeable on startup, where we refresh
the test tree of each project file individually.

refresh-test-tree can still take several seconds at startup, but now the
majority of that time appears to be in `(.publishTestTree client tree)`,
which is something to research later (it doesn't seem like it should be
slow).

In clj-kondo this patch cuts the initial refresh-test-tree time down
from about 4-5 seconds to 2 or less.

* Fix exception when testing label isn't a string

The first argument to clojure.test/testing is a string, but it can be
constructed via `(str "test thing " thing)`, or similar. The test-tree
code parses these nodes as sexps, but assumes that they're strings,
causing an exception. This fixes the exception.