Avoid coordinating around mutable var
In integration tests when a client makes a request we now wait for the
server's response to that specific request. Before we were waiting for a
request whose id happened to be the same as the id held in an atom.
That was fragile, and though I don't think it ever bit us, there's no
reason to do it.