Be sure to remove same request or notification as we return
Removes a potential source of raciness by avoiding clearing the wrong
stub data.

In integration tests, the "client" is really just a collection of stubs
that record the server's requests and its responses to the client's
requests. The tests wait for a particular message to be seen, then
remove that object from the stubs so the message doesn't pollute future
tests. The way we were doing this was to inspect the message's "method"
twice, once when finding the message (read) and again when removing it
from the stub (write). Much of the request/response cycle happens
asynchronously, and we weren't accounting for the possibility that the
server could make a second request using the same method between the
read and write. In that case, we could have removed a different message
from the one that we returned.

I don't think that was biting us, but it could in the future, and so
best to remove the problem now.