Integration test cleanup (#935)
* Print integration test log only once

* Add timestamp to integration test logs

* Log exceptions in un-dereffed futures

Otherwise, these exceptions are lost, because Clojure catches and holds
them, waiting for the future to be dereffed.

* Stop using unreliable tee helpers

See https://github.com/clojure-lsp/lsp4clj/pull/2 for details about how
these helpers are unreliable. See comments near Launcher/createLauncher
for a better way to trace I/O through lsp4j.

* Avoid coordinating around mutable var

In integration tests when a client makes a request we now wait for the
server's response to that specific request. Before we were waiting for a
request whose id happened to be the same as the id held in an atom.
That was fragile, and though I don't think it ever bit us, there's no
reason to do it.

* Be sure to remove same request or notification as we return

Removes a potential source of raciness by avoiding clearing the wrong
stub data.

In integration tests, the "client" is really just a collection of stubs
that record the server's requests and its responses to the client's
requests. The tests wait for a particular message to be seen, then
remove that object from the stubs so the message doesn't pollute future
tests. The way we were doing this was to inspect the message's "method"
twice, once when finding the message (read) and again when removing it
from the stub (write). Much of the request/response cycle happens
asynchronously, and we weren't accounting for the possibility that the
server could make a second request using the same method between the
read and write. In that case, we could have removed a different message
from the one that we returned.

I don't think that was biting us, but it could in the future, and so
best to remove the problem now.

* More gracefully shutdown client and server in integration tests

* Explaining variable

* Change arg order

Co-authored-by: Eric Dallo <ericdallo06@hotmail.com>