Skip var bodies when analyzing deps
This decreases uncached startup time by 60-70%. See
https://github.com/clj-kondo/clj-kondo/pull/1674.

This also triples the batch size of the dep analysis, since clj-kondo
seems to analyze about 3 times as many files in the same amount of time.
I don't have numbers on how much memory was being saved by batching
analysis, so that needs to be double-checked.

This also skips all linting, and analysis of var-usages for deps since
we don't use those. The performance gain is almost exclusively from
skipping var bodies, however, so that's the interesting part.