Fix drag for calva (#955)
* Wait for client to apply edit before moving cursor

This fixes a bug that was noticed in Calva. The server was sending both
workspace/applyEdit and window/showDocument simultaneously and
asynchronously. Sometimes Calva received and/or processed the
showDocument request before the applyEdit request, meaning the cursor
didn't end up in the correct place. To fix this we now wait for the
client to acknowledge the applyEdit command before sending the
showDocument command, ensuring they are processed in the correct order.

This also teaches the integration tests how to mock a client's response
to a server request.

* Ensure split comment nodes have correct position metadata

* Hold nodes, not zlocs when processing drag

* Edit only dragged clauses

* Cursor is at position, not range

* Fix whitespace

* Changelog

* Fix typo

* Record mock client request before sending it

* Fix integration test

* Remove unused private var

* Always wait for edits, so we never navigate too early

* Make client->server messages atomic

Now that we can send responses to asynchronous server requests, we
should ensure that each message from the client to the server is sent
atomically. This should avoid test flakiness, and may even fix old test
flakiness, though I don't see how.

* Touch tests

* Avoid repeated conversion of clauses to sequences of nodes

* Remove extraneous datastructure