Fix drag in 'are' when clojure.test is aliased (#968)
If clojure.test was aliased, for example to `test`, dragging within
`test.are` didn't behave as expected.

There were two possible fixes...

1. Look at the name of the symbol, not the whole thing.
2. Use the clj-kondo analysis to check whether the call site actually
   uses `clojure.test/are`.

The second option is technically more correct, but I picked the first.
The first option is (very slightly) faster, but that wasn't my main
motivator. Instead, I kind of like that we look only at the symbol for
`are`, along with `cond`, `case`, and other special functions, as well
as the binding forms `let`, etc. This means that other libraries that
define functions with these same names get treated the same. The
argument is that if a library author names a var `let`, its parameters
should have the same structure as `clojure.core/let`'s.

Fixes #967.