Replace quadratic algorithm with linear for semantic tokens
To find the prior token (while converting absolute tokens to relative),
the system was using `nth` on a seq, which is a linear time operation
(as opposed to on a vector, where it would be a constant time
operation). Since it was doing this in a loop, semantic token
calculation was a quadratic time operation overall.

This switches from using `nth`, to using `(partition 2 1)` to make
groups of the previous and subsequent token. This converts semantic
token calculation back to a linear time operation.

In clojure.core this reduces the time to calculate the full semantic
tokens from 7s to 0.3s, a 95% reduction in time.