Avoid going to clj def from cljs usage (#1003)
```clojure
;; a.cljs
(ns a)
(def f)
;; a.clj
(ns a)
(def f)
;; b.cljs
(ns b
  (:require [a]))
|a/f
```

In the above code, find-definition should always go to `a.cljs`, but it
could go to `a.clj`. Which one it picked was arbitrary, because both
definitions match.

This bug was accidentally introduced in
d5e1ef39bc17cd8bb7f94bf6c9f611335dd6b2af, which was developed to fix
https://github.com/clojure-lsp/clojure-lsp/issues/980. It might have
been caught earlier because there are tests around this. Unfortunately,
the tests only fail if the cljs file is loaded before the clj file,
which isn't how they were written originally.

This patch also addresses 980, but avoids the incorrect navigation.