Fix find-declaration in cljc (#1034)
Fix to ensure that when invoking find-declaration from a CLJ element in
a CLJC file, goes to the CLJ declaration, not a matching CLJS
declaration.

This functionality was present but was lost in a recent commit. Since
that hasn't been released yet, I haven't added a changelog.

This also backfills tests for finding the declaration of a var declared
with `:refer`, fixing #1025.