Timeout integration tests after waiting 30s for server response
Instead of letting the integration tests wait forever for a response
from the server, this times them out after 30 seconds. This will provide
some additional information in the logs about what the client was
waiting for, helping to debug flaky tests. If 30 seconds isn't long
enough for certain requests, we can either make the timeout longer, or
make the timeout configurable per request.