Avoid blocking in go blocks (#1064)
Avoids using blocking operations (the !! ones) within go blocks, as this
will block the limited number of threads dedicated to serving go blocsk.

In some cases we switch to using the parking (!) versions, when we can
make them visible from a `go`. In other cases, we switch to put! and
take!  which, unlike >!! and <!!, operate asynchronously. This means
they return immediately, after enqueing the channel operation, instead
of blocking until the operation is done.

Also changes the debug build so that we get early warning of this
problem.