Run integration tests without I/O buffering
By turning on tracing in lsp4j and additional logging in the integration
test client listener loop, it became apparent that sometimes lsp4j
thinks that it is sending data, but the client never receives it,
leading to intermittent test failures.

To be more precise, lsp4j generates a trace of a message [1], and writes
that trace [2] to a file. We can observe this far; the trace appears in
the file.

Then lsp4j passes [3] the message from the tracer to the thing that
actually writes to stdout [4]. But, the client listener, which is
reading from the server's stdout, never receives the message header [5].

This leaves two possibilities. Either lsp4j falls down somewhere between
tracing [2] and writing to stdout [4]. Or lsp4j does actually write, but
the message gets stuck in a buffer somewhere.

I made a guess that we're having the second problem, that messages are
stuck in a stdout buffer. This patch is an attempt to fix that problem.

With some searching [6] it turns out that this is a common problem for
programs that need to provide a "live" unbuffered stream to their
consumers. See here [7] for a brief description and common examples of
the problem and here [8] for a more in-depth explanation.

There are many possible fixes, but this patch uses the GNU coreutils
program `stdbuf` [9]. I'm not actually sure that stdbuf will have any
effect on a Java program. But, it's worth a shot.

[1]: https://github.com/eclipse/lsp4j/blob/ab3c4891699fd3b3bd057a13df31982fd5bb96b3/org.eclipse.lsp4j.jsonrpc/src/main/java/org/eclipse/lsp4j/jsonrpc/TracingMessageConsumer.java#L101
[2]: https://github.com/eclipse/lsp4j/blob/ab3c4891699fd3b3bd057a13df31982fd5bb96b3/org.eclipse.lsp4j.jsonrpc/src/main/java/org/eclipse/lsp4j/jsonrpc/TracingMessageConsumer.java#L109-L112
[3]: https://github.com/eclipse/lsp4j/blob/ab3c4891699fd3b3bd057a13df31982fd5bb96b3/org.eclipse.lsp4j.jsonrpc/src/main/java/org/eclipse/lsp4j/jsonrpc/TracingMessageConsumer.java#L114
[4]: https://github.com/eclipse/lsp4j/blob/ab3c4891699fd3b3bd057a13df31982fd5bb96b3/org.eclipse.lsp4j.jsonrpc/src/main/java/org/eclipse/lsp4j/jsonrpc/json/StreamMessageConsumer.java#L66-L70
[5]: https://github.com/clojure-lsp/clojure-lsp/blob/34d4765b0ef6a73beff8d97962301ec4e9c5f33e/cli/integration-test/integration/client.clj#L76
[6]: https://unix.stackexchange.com/questions/25372/turn-off-buffering-in-pipe
[7]: https://www.perkin.org.uk/posts/how-to-fix-stdio-buffering.html
[8]: https://www.pixelbeat.org/programming/stdio_buffering/
[9]: https://www.gnu.org/software/coreutils/manual/html_node/stdbuf-invocation.html